Test/fix the majority of ChanServ's DB code, make tests run clean.
* Add skeleton ChannelService to help get tests to run.
* Make last_activity for chanserv_channels required, and provide a
  default inside the initial registration. There's no API for this,
  currently, you'd just have to set it using Channel#last_activity=
  (and then calling Channel#save).
* Make Channel#[]= delete a flag if the value is nil or false.
* Fix a bunch of problems with Sequel stemming from the fact that I
  forgot associations aren't datasets, just arrays.
* Make ChannelService::Helper methods wait until the first time it's
  instantiated to spring into existence - to make sure that the
  constants they base themselves off of have been created.
* Fix a test for catching exceptions in core_test.rb.
* Tests now clean themselves unless the KEEP_TEST_DB environment var
  is set to 'yes'. Any other value will result in the test DB being
  wiped from the file system upon successful completion of the tests,
  assuming all the tests passed. If the tests errored
  catastrophically, then the test DB will remain.
* Test and fix everything in chanserv except for the succession of a
  successor to owner position when a founder leaves or is removed
  from power.

+++ test/teststrap.rb
@@ -10,20 +10,38 @@
 $LOAD_PATH.unshift File.expand_path('../',    File.dirname(__FILE__))
 $LOAD_PATH.unshift File.expand_path('../lib', File.dirname(__FILE__))
 
-# Require sequel, delete any old test db, and create the connection here
-require 'sequel'
-File.delete('db/test.db') rescue nil
-$db = Sequel.sqlite('db/test.db')
-
-# Run the migrations here, and then the models will have them to load
-Sequel.extension :migration
-Sequel::Migrator.run Sequel::Model.db, 'db/migrations'
-
 require 'rubygems'
 require 'kythera'
 require 'riot'
 #require 'riot/rr'
-Riot.reporter = Riot::VerboseStoryReporter
+
+# Set up Riot a little more explicitly
+Riot.verbose # use verbose error reporting (dump a callstack on errors)
+Riot.alone!  # require riot to be run manually, so we can do post-test cleanup
+
+# Require sequel, delete any old test db, and create the connection here. This
+# is done in a BEGIN block so it executes before the requires above.
+BEGIN {
+  require 'sequel'
+  File.delete('db/test.db') rescue nil
+  $db = Sequel.sqlite('db/test.db')
+
+  # Run the migrations here, and then the models will have them to load.
+  Sequel.extension :migration
+  Sequel::Migrator.run Sequel::Model.db, 'db/migrations'
+}
+
+# Run the tests after everything has been loaded up (variables defined etc), and
+# delete the test db afterwards (but only if KEEP_TEST_DB is not set).
+END {
+  status = Riot.run.success?
+
+  unless ENV['KEEP_TEST_DB'] == 'yes'
+    File.delete('db/test.db') rescue nil
+  end
+
+  exit(status)
+}
 
 # For all tests
 $config    = nil
